feat: add auth

+++ src/main/java/com/example/itemserver/controller/FolderController.java
@@ -2,6 +2,7 @@ package com.example.itemserver.controller;
 
 import com.example.itemserver.dto.ResponseDTO;
 import com.example.itemserver.entity.Folder;
+import com.example.itemserver.service.AuthorizationService;
 import com.example.itemserver.service.FolderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.*;
@@ -19,6 +20,8 @@ public class FolderController {
 
     @Autowired
     private RestTemplate restTemplate;
+    @Autowired
+    private AuthorizationService authorizationService;
 
     public static final String url = "https://userserver-production.up.railway.app";
 
@@ -27,7 +30,10 @@ public class FolderController {
     }
     @PostMapping("/add")
     public ResponseDTO addItem(@RequestPart("name") String name,
-                               @RequestPart("id") String id){
+                               @RequestPart("id") String id,
+                               @RequestHeader(name = "Authorization") String token){
+        if(!authorizationService.authorization(token)) return ResponseDTO.authFailed();
+
         ResponseDTO response = new ResponseDTO();
         response = folderService.addFolder(name, id);
         return response;
@@ -35,42 +41,59 @@ public class FolderController {
 
     @PostMapping("/updateFolder")
     public ResponseDTO updateFolder(@RequestPart("name") String name,
-                                    @RequestPart("newName") String newName){
+                                    @RequestPart("newName") String newName,
+                                    @RequestHeader(name = "Authorization") String token){
+        if(!authorizationService.authorization(token)) return ResponseDTO.authFailed();
+
         ResponseDTO response = new ResponseDTO();
         response = folderService.updateFolder(name,newName);
         return response;
     }
 
     @PostMapping("/findFolder")
-    public ResponseDTO findFolder(@RequestPart("name") String name){
+    public ResponseDTO findFolder(@RequestPart("name") String name,
+                                  @RequestHeader(name = "Authorization") String token){
+        if(!authorizationService.authorization(token)) return ResponseDTO.authFailed();
+
         ResponseDTO response = new ResponseDTO();
         response = folderService.findFolder(name);
         return response;
     }
 
     @PostMapping("/getAllFolderItem")
-    public ResponseDTO getAllFolderItem(@RequestPart("name") String name){
+    public ResponseDTO getAllFolderItem(@RequestPart("name") String name,
+                                        @RequestHeader(name = "Authorization") String token){
+        if(!authorizationService.authorization(token)) return ResponseDTO.authFailed();
+
         ResponseDTO response = new ResponseDTO();
         response = folderService.getAllFolderItem(name);
         return response;
     }
 
     @DeleteMapping(value = "/delete/{id}")
-    public ResponseDTO deleteFolder(@PathVariable(name = "id") Integer id) {
+    public ResponseDTO deleteFolder(@PathVariable(name = "id") String id,
+                                    @RequestHeader(name = "Authorization") String token) {
+        if(!authorizationService.authorization(token)) return ResponseDTO.authFailed();
+
         ResponseDTO response = new ResponseDTO();
         response = folderService.deleteFolder(id);
         return response;
     }
 
     @GetMapping("/list")
-    public ResponseDTO getAllItem(){
+    public ResponseDTO getAllItem(@RequestHeader(name = "Authorization") String token){
+        if(!authorizationService.authorization(token)) return ResponseDTO.authFailed();
+
         ResponseDTO response = new ResponseDTO();
         response = folderService.getAllFolder();
         return response;
     }
 
     @GetMapping(value = "/findByOwner/{id}")
-    public ResponseDTO findItem(@PathVariable(name = "id") int ownerId){
+    public ResponseDTO findItem(@PathVariable(name = "id") String ownerId,
+                                @RequestHeader(name = "Authorization") String token){
+        if(!authorizationService.authorization(token)) return ResponseDTO.authFailed();
+
         ResponseDTO response = new ResponseDTO();
         response = folderService.getAllFolderByOwner(ownerId);
         return response;
